4.2.0 beta and ALTB-344
BRATHL
- brathl Matrices clean-up
- other brathl clean-up, issued from ReadBratRecord analysis
- continued clean-up: uint32_t in matrices related code; other, such as iterators and names.

GUI
- raster custom raster file specification for maps
- WMS authentication bug (incomplete solution)
- NcDump dialog in qtc
[01256a3] NcDump dialog (debug versions only). IDEs: deploy netcdf command line executables.
- NcDump dialog resize adjustments; . NcDump dialog adjustments to mac.

Documentation and General Items
- Updated reference manual
- Updated SUM figures.
- Removed RC tag
- Added missing QGIS dependency (wms provider)
- External utilities deployed in mac
- deploy nefcdf utilities with qtc linux
- ALTB-359 note in README
- ALTB-359 note in source code comment with complementary extra-code solution

+++ process/BratProcess.h
@@ -532,10 +532,10 @@ namespace processes
 		static std::string PositionModeStr( CBratProcess::PositionMode mode );
 
 		static bool CheckPositionValue
-		( double		value,
-			double		min,
-			double		step,
-			uint32_t	count,
+		( double	value,
+			double	min,
+			double	step,
+			size_t	count,
 			CBratProcess::PositionMode	mode,
 			uint32_t	&position );
 
@@ -675,7 +675,7 @@ namespace processes
 		void MergeDataValue
 		( double* data,
 			double* values,
-			uint32_t nbValues,
+			size_t nbValues,
 			uint32_t indexExpr,
 			double* countValues,
             double* meanValues,
@@ -689,13 +689,7 @@ namespace processes
 		//	double* countValue,
         //	double* meanValue );
 
-		void FinalizeMergingOfDataValues
-		( double* data,
-			uint32_t indexExpr,
-			uint32_t nbValues,
-			double* countValues,
-            double* meanValues);
-
+		void FinalizeMergingOfDataValues( double* data, size_t indexExpr, size_t nbValues, double* countValues, double* meanValues);
 	};
 
 }
